Configure effort for org
We want to have effort up to 2 hours configured.  Two hours I consider a big
effort.  Actually I need low medium and high.  Using the numbers allows to use
filter of agenda.  To be able to find all big effort, I need to exclude tasks
for which no effort is set.  I want to minimize the effort to maintain the
"effort".